🚚 Move object arbitraries into /arbitrary
